chore: squash me

+++ gbpplanner-rs/src/ui/settings.rs
@@ -290,7 +290,7 @@ fn scale_ui(
             let scale_factor = match ui_state.scale_type {
                 UiScaleType::None => 1.0,
                 UiScaleType::Custom => ui_state.scale_percent as f32 / 100.0,
-                UiScaleType::Window => 1.0 / window.scale_factor() as f32,
+                UiScaleType::Window => 1.0 / window.scale_factor(),
             };
             egui_settings.scale_factor = scale_factor;
         }
